# PFTbackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/savings": {
            "get": {
                "tags": [
                    "Savings"
                ],
                "summary": "Get list of savings",
                "operationId": "a900b455832f140eff67adcf8ed7cebc",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by saving name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_target",
                        "in": "query",
                        "description": "Minimum target amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_target",
                        "in": "query",
                        "description": "Maximum target amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_current",
                        "in": "query",
                        "description": "Minimum current amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_current",
                        "in": "query",
                        "description": "Maximum current amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of savings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Emergency Fund"
                                                    },
                                                    "target_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1000
                                                    },
                                                    "current_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 500
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Savings"
                ],
                "summary": "Create a new saving goal",
                "operationId": "f4a721aff1433a0e2a440f98bde97240",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "target_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Emergency Fund"
                                    },
                                    "target_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1000
                                    },
                                    "current_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saving goal created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Emergency Fund"
                                        },
                                        "target_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000
                                        },
                                        "current_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/savings/{saving}": {
            "get": {
                "tags": [
                    "Savings"
                ],
                "summary": "Get a specific saving goal",
                "operationId": "80e4a011bb184d4e718a8d1255d2f7f0",
                "parameters": [
                    {
                        "name": "saving",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saving goal details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Emergency Fund"
                                        },
                                        "target_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000
                                        },
                                        "current_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 500
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Saving goal not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Savings"
                ],
                "summary": "Update a saving goal",
                "operationId": "627eedc3f48e29c71e141805d257bef5",
                "parameters": [
                    {
                        "name": "saving",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "target_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Emergency Fund"
                                    },
                                    "target_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1000
                                    },
                                    "current_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saving goal updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Emergency Fund"
                                        },
                                        "target_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000
                                        },
                                        "current_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 500
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Savings"
                ],
                "summary": "Delete a saving goal",
                "operationId": "a5419fc7bbf5291c94b883f71a286a99",
                "parameters": [
                    {
                        "name": "saving",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saving goal deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Saving deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}